* main: (30 commits)
  Update RelayService.swift
  Update RelayView.swift
  Update RelayView.swift
  Update CHANGELOG.md
  On relay remove, close all subs, the disconnect and delete socket.
  Copy UserId from Profile View
  Fix swiftlint errors
  Update changelog
  Move event parsing to a background thread
  Fix replies queries
  Update RelayService.swift
  Fixed bug. Better logging.
  Fix tests
  Add test for parsing stub events
  Added NoteOptionsButton to NosTests target for unit tests
  Hack to fix confirmationDialog not showing up
  Better logging and bug fix.
  Copy Note Identifier works
  Changelog. Lint.
  Capped event send attempts to 5
  ...

+++ Nos/Views/RelayView.swift
@@ -9,8 +9,8 @@ import SwiftUI
 import CoreData
 
 struct RelayView: View {
-    
     @Environment(\.managedObjectContext) private var viewContext
+    @EnvironmentObject private var relayService: RelayService
     
     @State var newRelayAddress: String = ""
     
@@ -29,8 +29,20 @@ struct RelayView: View {
                     .onDelete { indexes in
                         for index in indexes {
                             let relay = relays[index]
+
+                            guard let address = relay.address else { continue }
+
+                            if let socket = relayService.socket(for: address) {
+                                for subId in relayService.activeSubscriptions {
+                                    relayService.sendClose(from: socket, subscription: subId)
+                                }
+
+                                relayService.close(socket: socket)
+                            }
+                            
                             viewContext.delete(relay)
                         }
+                        try! viewContext.save()
                     }
                     if relays.isEmpty {
                         Localized.noRelaysMessage.view
@@ -78,6 +90,7 @@ struct RelayView: View {
             
             let relay = Relay(context: viewContext)
             relay.address = newRelayAddress.lowercased().trimmingCharacters(in: .whitespacesAndNewlines)
+            relay.createdAt = Date.now
             newRelayAddress = ""
 
             do {
